requirements.txt
flask
python-dotenv
requests
gunicorn
google-cloud-storage
google-cloud-firestore
google-cloud-logging